Added options to the constructor
When using serverless endpoints, you sometimes have access to a specific folder when write access is needed.  In the case of Webtask (webtask.io), this folder is /tmp/.  Because the current code assumes the active directory is the one to be used by NodeJs to store and retrieve temporary files, this wouldn't work.  
I added a "options" parameter to the constructor to pass in this tempFilePath option so that it works in my context.  
I understand that there is a gazillion ways to do this and I don't know your coding standards so I tried to do changes that would be backward compatible.  Up to you to see how you would rather implement this.
This is more of a way to show what I needed for this project to work than an actual PR.